Refactored RSpec support kit temp dir setup.

+++ spec/support/kit/manage_temp_dir.rb
@@ -1,12 +1,17 @@
+require "rspec/core/shared_context"
+
+module RSpecSupportKitTempDirContext
+  extend RSpec::Core::SharedContext
+  let (:temp_dir) { File.expand_path "../../../../tmp/rspec", __FILE__ }
+end
+
 RSpec.configure do |config|
-  temp_dir = File.expand_path "../../../../tmp/rspec", __FILE__
+  config.include RSpecSupportKitTempDirContext
 
-  config.before :suite do
-    FileUtils.rm_rf(temp_dir) if File.exist?(temp_dir)
-    FileUtils.mkdir_p(temp_dir) unless File.exist?(temp_dir)
-  end
-
-  config.after :suite do
-    FileUtils.rm_rf(temp_dir) if File.exist?(temp_dir)
+  config.before do |example|
+    if example.metadata[:temp_dir]
+      FileUtils.rm_rf(temp_dir) if File.exist?(temp_dir)
+      FileUtils.mkdir_p(temp_dir)
+    end
   end
 end
